fix: Some name mistakes solved

+++ MarketApi/Models/DiscountCode.cs
@@ -6,6 +6,6 @@ namespace MarketApi.models
     {
         [Key] public int ID { get; set; }
         [Required] public string Code { get; set; }
-        [Required] public decimal DiscountPrecent { get; set; }
+        [Required] public int DiscountPrecent { get; set; }
     }
 }
